More resilient importer

+++ importer.rb
@@ -8,11 +8,13 @@ class Importer
   def self.download_historical_data
     puts 'Importing ...'
     csv_contents = create_ticker_list
-    i = 0
-    csv_contents.each do |row|
-      download_historical_stock_data row[0].split('|')[0]
-      i += 1
-      break if i > 10 # TODO: Figure out how to download all data in a way that doesn't over-tax Yahoo's API
+    csv_contents.each_with_index do |row, i|
+      begin
+        ticker = row[0].split('|')[0]
+        download_historical_stock_data(ticker)
+      rescue 
+        puts "Unable to import ticker: #{ticker}"
+      end
     end
   end
 
@@ -26,12 +28,14 @@ class Importer
   end
 
   def self.download_historical_stock_data(ticker)
+    file_path = "imported/#{ticker}.csv"
+    return if File.exist?(file_path)
+
     csv = YQuotes::Yahoo.new.get_csv(ticker, '2010-01-01', '2019-12-31', 'd')
 
     #noinspection RubyArgCount
-    FileUtils.mkdir('imported') unless File.directory?('imported')
+    FileUtils.mkdir_p('imported') unless File.directory?('imported')
 
-    file_path = "imported/#{ticker}.csv"
 
     CSV.open(file_path, 'w') do |out|
       csv.each do |row|
